requirements.txt
alembic==1.11.1
amqp==5.1.1
annotated-types==0.5.0
async-timeout==4.0.2
Babel==2.12.1
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
cffi==1.15.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
config==0.5.1
cryptography==41.0.2
Flask==2.3.2
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-JSONRPC==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-MySQLdb==1.0.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
Mako==1.2.4
MarkupSafe==2.1.3
mypy==1.4.1
mypy-extensions==1.0.0
mysqlclient==2.2.0
passlib==1.7.4
prompt-toolkit==3.0.39
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydentic==0.0.1.dev3
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-stdnum==1.18
redis==4.6.0
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.19
sqlalchemy-stubs==0.4
tomli==2.0.1
typeguard==2.13.3
typing-inspect==0.8.0
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.16.2

